# txt: use Data.Text.Text based implementation