Use `Data.Text.Text` to improve the performance of the `txt` widget.

On my workload that makes heavy use of `txt` this substantially
reduces allocation and somewhat reduces CPU time.  Total
CPU/allocation and top cost centres follow.

Before:

```
total time  =        0.62 secs   (623 ticks @ 1000 us, 1 processor)
total alloc = 769,766,368 bytes  (excludes profiling overheads)

COST CENTRE         MODULE                               SRC                                                  %time %alloc

str                 Brick.Widgets.Core                   src/Brick/Widgets/Core.hs:(341,1)-(356,70)            32.4   38.7
safeWcswidth        Graphics.Text.Width                  src/Graphics/Text/Width.hs:54:1-51                     9.3    3.1
safeWcwidth         Graphics.Text.Width                  src/Graphics/Text/Width.hs:49:1-29                     5.8    6.2
vBox                Brick.Widgets.Core                   src/Brick/Widgets/Core.hs:(481,1)-(483,41)             5.6   11.3
txt                 Brick.Widgets.Core                   src/Brick/Widgets/Core.hs:367:1-20                     4.5   14.2
…
```

After:

```
total time  =        0.54 secs   (542 ticks @ 1000 us, 1 processor)
total alloc = 432,922,048 bytes  (excludes profiling overheads)

COST CENTRE         MODULE                               SRC                                                  %time %alloc

safeWctwidth        Graphics.Text.Width                  src/Graphics/Text/Width.hs:59:1-53                    14.8   10.9
txt                 Brick.Widgets.Core                   src/Brick/Widgets/Core.hs:(367,1)-(381,63)            14.6   10.8
safeWcwidth         Graphics.Text.Width                  src/Graphics/Text/Width.hs:49:1-29                    10.9   11.0
vBox                Brick.Widgets.Core                   src/Brick/Widgets/Core.hs:(514,1)-(516,41)             9.0   20.1
cropResultToContext Brick.Widgets.Internal               src/Brick/Widgets/Internal.hs:(82,1)-(87,47)           4.6    3.8
…
```