@frasertweedale 2022-10-16
Full profiling output:

- [`master`](https://github.com/jtdaugherty/brick/files/9793836/txt-0.prof.txt)
- [This PR](https://github.com/jtdaugherty/brick/files/9793837/txt-1-strict.prof.txt)
- [This PR, converted to lazy Text](https://github.com/jtdaugherty/brick/files/9793838/txt-2-lazy.prof.txt)