@frasertweedale 2022-10-17
@jtdaugherty I do wonder if it makes sense to migrate `str` to pack the string and invoke `txt` now.  It doesn't make much difference for *purebred* as we don't use `str` much.  Which also means *purebred* is not suitable for benchmarking the change either.  But I'll leave it for you to consider, as it's a trivial code change.