@jtdaugherty 2022-10-17
Yeah, I had the same thought, if for no other reason than to reduce code duplication. At this point in everyday Haskell usage I hope it's a safe assumption that `txt` will get used more than `str` anyway.