@jtdaugherty 2022-10-23
This change is now released in `1.4`.